WeatherApp/requirements.txt
cassandra-driver==3.25.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
configparser==5.3.0
confluent-kafka==2.0.2
contourpy==1.0.7
cycler==0.11.0
daterangestr==0.0.3
decorator==5.1.1
experta==1.9.4
fonttools==4.39.2
frozendict==1.2
future==0.18.3
geocoder==1.38.1
geojson==2.5.0
geomet==0.2.1.post1
idna==3.4
kafka-python==2.0.2
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
openweather==0.11
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyowm==3.3.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
ratelim==0.1.6
requests==2.28.2
schema==0.6.7
scipy==1.10.1
six==1.16.0
spicy==0.16.0
urllib3==1.26.15
